Add Optional tiles to analysis
