+++ src/main/scala/org/globalforestwatch/treecoverloss/TreeLossSummary.scala
@@ -20,14 +20,15 @@ object TreeLossSummary {
     new CellVisitor[Raster[TreeLossTile], TreeLossSummary] {
       def register(raster: Raster[TreeLossTile], col: Int, row: Int, acc: TreeLossSummary): TreeLossSummary = {
 
-
         // This is a pixel by pixel operation
         val loss: Int = raster.tile.loss.get(col, row)
         val gain: Int = raster.tile.gain.get(col, row)
         val tcd2000: Int = raster.tile.tcd2000.get(col, row)
         val tcd2010: Int = raster.tile.tcd2010.get(col, row)
-        val co2Pixel: Double = raster.tile.co2Pixel.getDouble(col, row)
-        val gadm36: Int = raster.tile.gadm36.get(col, row)
+
+        // If we don't have these tiles use default values for pixel
+        val co2Pixel: Double = raster.tile.co2Pixel.map(_.getDouble(col, row)).getOrElse(0)
+        val gadm36: Int = raster.tile.gadm36.map(_.get(col, row)).getOrElse(0)
 
         val tcd2000Thresh: Int = TreeCoverDensity.threshold(tcd2000)
         val tcd2010Thresh: Int = TreeCoverDensity.threshold(tcd2010)
